requirements.txt
Flask==3.1.0
Werkzeug==3.1.3
Jinja2==3.1.6
MarkupSafe==3.0.2
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.40
PyMySQL==1.1.1
greenlet==3.2.2
Flask-Migrate
python-dotenv==1.1.0
blinker==1.9.0
itsdangerous==2.2.0
click==8.1.8
typing_extensions==4.13.2
cffi==1.17.1
cryptography==44.0.3
pycparser==2.22
beautifulsoup4 
schedule
requests
Flask-APScheduler==1.13.1
PyJWT
Flask-Testing
JWTManager
Flask-JWT-Extended==4.4.4
Flask-Login==0.6.3